+++ src/events/misc.h
@@ -7,7 +7,7 @@ extern "C" {
 
 void event_allaround_extract_find_colon   (struct irc_message_compo *);
 void event_allaround_extract_remove_colon (struct irc_message_compo *);
-void event_bounce                         (struct irc_message_compo *);
+void event_serverFeatures                 (struct irc_message_compo *);
 void event_channelCreatedWhen             (struct irc_message_compo *);
 void event_channelModeIs                  (struct irc_message_compo *);
 void event_channel_forward                (struct irc_message_compo *);