+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased] ##
+### Changed ###
+- `RPL_BOUNCE`->`RPL_ISUPPORT`
+
 ### Fixed ###
 - Reported by [MAGA](https://github.com/MakeItGreatAgain):
   - Reproducible crash from protocol fuzzing